+++ CHANGES.md
@@ -10,6 +10,10 @@ Unreleased
 * atdgen: Fix JSON I/O for inline records (#419)
 * atdgen: The deprecated `atdgen` library is no longer available.
           Use `atdgen-runtime` instead (#421)
+* atddiff: To ensure that all the intended types are checked when new
+           type definitions are added to an ATD file, an `--ignore`
+           option is now available. It makes atddiff fail if some types
+           are neither selected via `--types` nor ignored via `--ignore`.
 
 2.15.0 (2023-10-26)
 -------------------